requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
openai
python-multipart
requests
python-dotenv
aiofiles
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
aiohttp
asyncpg
jinja2
python-jose[cryptography]
passlib[bcrypt]
reportlab
mercadopago
